Пакеты обозревателя

+++ Server/Network/Packets/Explorer/ExplorerCreateSignFilePacket.cs
@@ -1,4 +1,5 @@
 ﻿using Publisher.Basic;
+using SocketCore.Extensions.Buffer;
 using SocketCore.Utils;
 using SocketCore.Utils.Buffer;
 
@@ -15,7 +16,7 @@ namespace Publisher.Server.Network.Packets.Explorer
 
         private static void Send(PublisherNetworkClient client, ExplorerActionResultEnum result)
         {
-            client.Network.Send((byte)PublisherClientPackets.ExplorerCreateSignFileResult, (byte)result);
+            client.Network.Send(PublisherClientPackets.ExplorerCreateSignFileResult, (byte)result);
         }
     }
 }
